+++ src/routes.js
@@ -205,3 +205,6 @@ export const isSearchUrl = (path) =>
 
 export const peopleUrl = community =>
   `${community ? '/c/' + community.slug : ''}/people`
+
+export const checklistUrl = community =>
+  `/c/${community.slug}?checklist=true`
